pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-chronicle"
version = "0.1.0"
description = "Historical Canadian Building Code Search"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "IskanderBlue"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Django core
    "Django>=5.0,<6.0",
    "django-ninja>=1.0",
    "psycopg2-binary>=2.9",
    
    # Building code search
    "building-code-mcp>=1.2.0",
    
    # LLM
    "anthropic>=0.18",
    
    # AWS
    "boto3>=1.34",
    
    # Auth & Payments
    "django-allauth>=0.60",
    "dj-stripe>=2.8",
    
    # Building code search extras
    "pymupdf>=1.23.0",
    "rapidfuzz>=3.5.0",
    "markdown>=3.6",
    
    # Utilities
    "python-dotenv>=1.0",
    "dj-database-url>=2.1.0",
    "gunicorn>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.7",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "pre-commit>=3.6",
]

[project.urls]
Homepage = "https://github.com/IskanderBlue/CodeChronicle"

[tool.setuptools.packages.find]
where = ["."]
include = ["code_chronicle*", "api*", "core*", "config*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line length handled separately

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "code_chronicle.settings.development"
python_files = ["test_*.py"]
addopts = "-v --tb=short"
